requirements.txt
# Core Dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0

# PostgreSQL Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Encryption
cryptography==41.0.7

# Google Sheets Integration (for questions)
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Scheduling
APScheduler==3.10.4

# Data Handling
pytz==2023.3

# Utilities
python-dateutil==2.8.2
